Create a new R Package

create_package() mimics usethis::create_package(), but uses defaults that I find more helpful. Specific actions can be found in the Details.

Usage

create_package(
  path,
  license = "mit",
  renv = TRUE,
  github = TRUE,
  pkgdown = github,
  coc = github,
  fledge = FALSE,
  cran_badge = fledge,
  open = interactive(),
  creator = NULL,
  ...
)

Arguments

path

A path. If it exists, it is used. If it does not exist, it is created, provided that the parent path exists.

license

[string] The license to use; corresponds to one of the usethis::use_*_license() functions

renv

[flag] Should {renv} be initialized for this package? This ensures that package dependencies are portable and reproducible and is recommended.

github

[flag] Should a Github repository be initialized for this package? The default is TRUE but will only work if you have set up your environment to work with usethis::use_github().

pkgdown

[flag] Should a {pkgdown} site be created for this package? Defaults to the value of github, so that publicly available packages get a site by default.

coc

[flag] Should a Code of Conduct be added to the package? Defaults to the value of github, so that publicly available packages get a Code of Conduct by default.

fledge

[flag] Should {fledge} be used for package NEWS and versions? The default is FALSE, since I only use NEWS and version tags for packages that I plan to distribute.

cran_badge

[flag] Should a badge indicating CRAN status be added to the README? Defaults to the value of fledge, since both are mainly used for packages that I plan to distribute.

open

[flag] Should the package open in a new RStudio window once setup is complete?

creator

[string] The name of the package creator (aka you). This is used to fill in the Authors field in the package Description, as well as the contact for the Code of Conduct (if used) and the copyright holder of the license (if applicable).

...

Additional arguments to pass to the license function

Value

invisible NULL